+++ transfer.py
@@ -3,12 +3,13 @@ import shutil
 
 """The 'transfer' command."""
 
-def transfer(installation, target):
+def transfer(installation, target, silent=None):
     """Copy savefiles between installations.
     
     Arguments:
     installation -- the installation to copy savefiles from
     target -- the installation to copy savefiles to
+    silent -- whether to ask before overwriting existing savefiles
     """
     if os.path.isdir('./'+installation):
         if os.path.isdir('./'+target):
@@ -18,8 +19,18 @@ def transfer(installation, target):
                     for file in os.scandir('./'+installation+'/data'):
                             if file.path.endswith(('.dat', '.dir', '.bak', '.checksum')):
                                 try:
-                                    shutil.copyfile(file.path, './'+target+'/data/'+file.name)
-                                    count += 1
+                                    if silent == 'silent':
+                                        shutil.copyfile(file.path, './'+target+'/data/'+file.name)
+                                        count += 1
+                                    else:
+                                        if not os.path.isfile('./'+target+'/data/'+file.name):
+                                            shutil.copyfile(file.path, './'+target+'/data/'+file.name)
+                                            count += 1
+                                        else:
+                                            overwrite = input(f"{file.name} already exists in {target}! Do you want to overwrite it (y/n)? ")
+                                            if overwrite.lower() == 'y':
+                                                shutil.copyfile(file.path, './'+target+'/data/'+file.name)
+                                                count += 1
                                 except OSError:
                                     print(f"Could not copy {file.name}")
                     print(f"Successfully copied {count//4} savefile(s).")
